feat(core): expose message and strict typing
