+++ tsconfig.json
@@ -6,6 +6,8 @@
     "declaration": true,
     "outDir": "dist",
     "strict": true,
+    "noImplicitAny": true,
+    "strictNullChecks": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
     "types": ["node"]
