+++ types/virtual-i18n-messages.d.ts
@@ -1,4 +1,5 @@
 declare module 'virtual:i18n-messages' {
-  const messages: Record<string, Record<string, Record<string, string>>>;
+  import type { I18nMessages } from 'vite-i18n-by-design';
+  const messages: I18nMessages;
   export default messages;
 }
